Create test file for Pass state handler
