+++ __tests__/StateMachine.test.ts
@@ -8,122 +8,6 @@ afterEach(() => {
 });
 
 describe('State Machine', () => {
-  describe('Pass State', () => {
-    test('should return initial input as result if neither `Result` nor `ResultPath` are specified', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            End: true,
-          },
-        },
-      };
-      const input = {
-        prop1: 'test',
-        prop2: 12345,
-      };
-
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(result).toEqual({
-        prop1: 'test',
-        prop2: 12345,
-      });
-    });
-
-    test('should return value of `Result` as result if `ResultPath` is not specified', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            Result: {
-              result1: 'result1',
-              result2: 67890,
-            },
-            End: true,
-          },
-        },
-      };
-      const input = {
-        prop1: 'test',
-        prop2: 12345,
-      };
-
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(result).toEqual({
-        result1: 'result1',
-        result2: 67890,
-      });
-    });
-
-    test('should return concatenation of initial input with itself if `ResultPath` is specified but `Result` is not specified', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            ResultPath: '$.concatenated',
-            End: true,
-          },
-        },
-      };
-      const input = {
-        prop1: 'test',
-        prop2: 12345,
-      };
-
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(result).toEqual({
-        prop1: 'test',
-        prop2: 12345,
-        concatenated: {
-          prop1: 'test',
-          prop2: 12345,
-        },
-      });
-    });
-
-    test('should return concatenation of initial input with `Result` if `ResultPath` is specified', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            ResultPath: '$.concatenated',
-            Result: {
-              result1: 'result1',
-              result2: 67890,
-            },
-            End: true,
-          },
-        },
-      };
-      const input = {
-        prop1: 'test',
-        prop2: 12345,
-      };
-
-      const stateMachine = new StateMachine(definition);
-      const result = await stateMachine.run(input);
-
-      expect(result).toEqual({
-        prop1: 'test',
-        prop2: 12345,
-        concatenated: {
-          result1: 'result1',
-          result2: 67890,
-        },
-      });
-    });
-  });
-
   describe('Wait State', () => {
     const mockSleepFunction = jest.fn();
     const mockDateNowFunction = jest.fn(() => 1670198400000); // 2022-12-05T00:00:00Z
